feat(tree): fix imports for ESM compatibility, rename api surface to use "block"
having "block" and "node" to refer to the same thing becomes confusing and we
should change this before the project is used.  note: there are some warnings
about imports ignored due to the no side effects policy on the forten modules.
But it seems to work fine.

BREAKING CHANGE:
API surface has changed, consult TreeConfig and TreeSettings